Move FullScreenAlarmNavHost to different package and modify transitions. Modify PostAlarmConfirmationScreen font.
